+++ src/Component/Dto/Actions/GameRestoreActionDto.php
@@ -1,6 +1,5 @@
 ﻿<?php namespace App\Component\Dto\Actions;
 
-use Doctrine\Common\Collections\Collection;
 use App\Component\Dto\GameDto;
 use App\Component\Type\PlayerColor;
 
@@ -13,5 +12,5 @@ class GameRestoreActionDto extends ActionDto
     
     public GameDto $game;
     public PlayerColor $color;
-    public Collection $dices; // DiceDto[]
+    public array $dices; // DiceDto[]
 }