+++ src/Component/Dto/Actions/MovesMadeActionDto.php
@@ -1,6 +1,4 @@
-﻿<?php namespace App\Component\Dto\Actions;
-
-use Doctrine\Common\Collections\Collection;
+<?php namespace App\Component\Dto\Actions;
 
 class MovesMadeActionDto extends ActionDto
 {
@@ -8,6 +6,6 @@ class MovesMadeActionDto extends ActionDto
     {
         $this->actionName = ActionNames::movesMade->value;
     }
-
-    public Collection $moves; // MoveDto[]
+    
+    public array $moves; // MoveDto[]
 }