+++ src/Component/Rules/Backgammon/Point.php
@@ -5,11 +5,11 @@ use Doctrine\Common\Collections\Collection;
 
 use App\Component\Type\PlayerColor;
 
-/// <summary>
-/// Represents one of 24 a triangles where a checker can be placed.
-/// 1 to 24.
-/// The bar is 0. Beared off is 25.
-/// </summary>
+/**
+ * Represents one of 24 a triangles where a checker can be placed.
+ * 1 to 24.
+ * The bar is 0. Beared off is 25.
+ */
 class Point
 {
     /** @var Collection | Checker[] */
@@ -28,13 +28,17 @@ class Point
     
     public function IsOpen( PlayerColor $myColor ): bool
     {
+        if ( $this->Checkers->isEmpty() ) {
+            return false;
+        }
+        
         //Opponent has less than two checkers on the point.
         //My own home is always open.
         return $this->Checkers->filter(
             function( $entry ) use ( $myColor ) {
                 return $entry->Color != $myColor;
             }
-            )->count() < 2 || $this->GetNumber( $myColor ) == 25;
+        )->count() < 2 || $this->GetNumber( $myColor ) == 25;
     }
     
     public function GetNumber( PlayerColor $player ): int
@@ -42,11 +46,11 @@ class Point
         return $player == PlayerColor::Black ? $this->BlackNumber : $this->WhiteNumber;
     }
     
-    //     public function __toString(): string
-    //     {
-    //         $color = $this->Checkers->count() ? $this->Checkers->First()->Color : "";
-    //         return "{$this->Checkers->count()} {$color} WN = {$this->WhiteNumber}, BN = {$this->BlackNumber}, ";
-    //     }
+    public function __toString(): string
+    {
+        $color = $this->Checkers->count() ? strval( $this->Checkers->first()->Color ) : "";
+        return "{$this->Checkers->count()} {$color} WN = {$this->WhiteNumber}, BN = {$this->BlackNumber}, ";
+    }
     
     public function IsHome( PlayerColor $player ): bool
     {